requirements-frozen.txt
amqp==5.1.1
async-timeout==4.0.2
attrs==21.4.0
aws-cdk-lib==2.46.0
billiard==3.6.4.0
blessed==1.19.1
boto3==1.24.94
botocore==1.27.94
cattrs==22.2.0
celery==5.2.7
chalice==1.27.3
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constructs==10.1.134
Deprecated==1.2.13
exceptiongroup==1.0.0rc9
inquirer==2.10.0
jmespath==1.0.1
jsii==1.70.0
kombu==5.2.4
packaging==21.3
prompt-toolkit==3.0.31
psycopg2-binary==2.9.4
publication==0.0.3
pydantic==1.10.2
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2022.5
PyYAML==6.0
readchar==4.0.3
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==1.3.24
typeguard==2.13.3
typing_extensions==4.4.0
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
